Fixed WinDirStat's completely broken visualisation
The bug is caused by the fact that an unsigned long long minus an
unsigned long long is ALWAYS positive. The quicksort function algorithm used
`signum` to compare unsigned long longs. With signum
always returning either 1 or 0 (because always positive), the sorting was
pretty much useless. With an arbitrary sorting order in multiple locations of
the program, nothing was grouped correctly and the wrong file extensions got a
colour assigned.

The bug was introduced in revision 151.